% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{check_address_parameters}
\alias{check_address_parameters}
\title{preemptive check for address-similarity request parameter errors}
\usage{
check_address_parameters(parameters)
}
\arguments{
\item{parameters}{- parameters list to be passed to specified Rosette API
endpoint}
}
\value{
Returns list of verified parameters to be sent to Rosette API
}
\description{
preemptive check for address-similarity request parameter errors
}
