% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{check_for_multipart}
\alias{check_for_multipart}
\title{Helper to check for file submission}
\usage{
check_for_multipart(parameters)
}
\arguments{
\item{parameters}{- JSON parameters}
}
\value{
true if multipart
}
\description{
Helper to check for file submission
}
