% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{check_morphology}
\alias{check_morphology}
\title{determine which morphology endpoint to use}
\usage{
check_morphology(parameters, endpoint)
}
\arguments{
\item{parameters}{- parameters list to be passed to specified Rosette API
endpoint}

\item{endpoint}{- Rosette API endpoint to be utilized}
}
\value{
Returns the specified morphology function
}
\description{
determine which morphology endpoint to use
}
