% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{create_multipart}
\alias{create_multipart}
\title{create a multipart}
\usage{
create_multipart(parameters)
}
\arguments{
\item{parameters}{- parameters list to be passed to specified Rosette API
endpoint}
}
\value{
Returns a multipart
}
\description{
create a multipart
}
