% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{error_check}
\alias{error_check}
\title{check if Rosette API response includes and error message}
\usage{
error_check(response)
}
\arguments{
\item{response}{- response from Rosette API}
}
\value{
Returns an error if one exists or the Rosette API response
}
\description{
check if Rosette API response includes and error message
}
