% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{get_headers}
\alias{get_headers}
\title{Helper to combine custom headers with the default ones}
\usage{
get_headers(user_key, content_type = "application/json", custom_headers = NULL)
}
\arguments{
\item{user_key}{- Rosette API authentication key}

\item{content_type}{- Header Content-Type}

\item{custom_headers}{- custom headers for the Rosette API}
}
\value{
Returns the combined h
}
\description{
Helper to combine custom headers with the default ones
}
