% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{post_endpoint}
\alias{post_endpoint}
\title{httr::POST request to specified Rosette API endpoint}
\usage{
post_endpoint(
  user_key,
  parameters,
  endpoint,
  url,
  custom_headers = NULL,
  url_parameters = NULL
)
}
\arguments{
\item{user_key}{- Rosette API authentication key}

\item{parameters}{- parameters list to be passed to specified Rosette API
endpoint}

\item{endpoint}{- Rosette API endpoint to be utilized}

\item{url}{- url for Rosette Api}

\item{custom_headers}{- custom headers for Rosette Api}

\item{url_parameters}{- query parameters}
}
\value{
Returns the Rosette API response
}
\description{
httr::POST request to specified Rosette API endpoint
}
