% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{serialize_address_params}
\alias{serialize_address_params}
\title{serialize address-similarity parameters}
\usage{
serialize_address_params(parameters)
}
\arguments{
\item{parameters}{- parameters list to be passed to address-similarity}
}
\value{
Returns the serialized parameters for the Rosette API
}
\description{
serialize address-similarity parameters
}
