% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{serialize_name_deduplication_parameters}
\alias{serialize_name_deduplication_parameters}
\title{serialize Rosette API parameters}
\usage{
serialize_name_deduplication_parameters(parameters)
}
\arguments{
\item{parameters}{- parameters list to be passed to name-deduplication}
}
\value{
Returns the serialized parameters for the Rosette API
}
\description{
serialize Rosette API parameters
}
