% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{serialize_name_parameters}
\alias{serialize_name_parameters}
\title{serialize Rosette API parameters}
\usage{
serialize_name_parameters(parameters)
}
\arguments{
\item{parameters}{- parameters list to be passed to either name-translation
or name-similarity}
}
\value{
Returns the serialized parameters for the Rosette API
}
\description{
serialize Rosette API parameters
}
