% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{serialize_parameters}
\alias{serialize_parameters}
\title{serialize Rosette API parameters}
\usage{
serialize_parameters(parameters)
}
\arguments{
\item{parameters}{- parameters list to be passed to specified Rosette API
endpoint}
}
\value{
Returns the serialized parameters for the Rosette API
}
\description{
serialize Rosette API parameters
}
