% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Api.R
\name{to_json}
\alias{to_json}
\title{convert Rosette API response content and headers to json}
\usage{
to_json(response)
}
\arguments{
\item{response}{- response from Rosette API}
}
\value{
Returns a vector containing json of Rosette API response and response headers
}
\description{
convert Rosette API response content and headers to json
}

